% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mcri.palettes}
\alias{mcri.palettes}
\title{Complete list of MCRI palettes}
\format{An object of class \code{list} of length 1.}
\usage{
mcri.palettes
}
\description{
Use \code{\link{mcriPalette}} to construct palettes of desired length.
}
\keyword{datasets}

