% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mcriPalette}
\alias{mcriPalette}
\title{A MCRI palette generator}
\usage{
mcriPalette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{MCRI}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of colour palettes based on the MCRI branding.
}
\examples{
mcriPalette("symbol")

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- mcriPalette(21, name = "symbol", type = "continuous")
image(volcano, col = pal)

}
\keyword{colors}
\keyword{colours,}
